projects/poller/pyproject.toml
[project]
name = "poller"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Samuel Hinton", email = "samuelreay@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.11",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "opentelemetry-instrumentation-httpx>=0.51b0",
    "prometheus-fastapi-instrumentator>=7.0.2",
    "uvicorn>=0.34.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/poller"]
